u32

+++ studio/editor_core/src/position.rs
@@ -93,12 +93,12 @@ impl Add<Size> for Position {
         if other.line == 0 {
             Position {
                 line: self.line,
-                column: self.column + other.column,
+                column: self.column + other.column as usize,
             }
         } else {
             Position {
-                line: self.line + other.line,
-                column: other.column,
+                line: self.line + other.line as usize,
+                column: other.column as usize,
             }
         }
     }
@@ -121,12 +121,12 @@ impl Sub for Position {
         if self.line == other.line {
             Size {
                 line: 0,
-                column: self.column - other.column,
+                column: self.column as u32 - other.column as u32,
             }
         } else {
             Size {
-                line: self.line - other.line,
-                column: self.column,
+                line: self.line as u32 - other.line as u32,
+                column: self.column as u32,
             }
         }
     }
